Add crate for shared test utility functions
See https://stackoverflow.com/a/44545091/1165155
