+++ .gitignore
@@ -1,4 +1,5 @@
 /target
+/test-utils/target
 
 # Logs
 logs
