+++ Cargo.toml
@@ -26,4 +26,4 @@ sqlx = { version = "0.5.0", features = ["chrono", "offline", "json", "mysql", "r
 thiserror = "1.0.0"
 
 [dev-dependencies]
-rand = "0.8.0"
+test-utils = { path="test-utils" }
