+++ src/notification/model.rs
@@ -260,7 +260,8 @@ mod tests {
     };
     use crate::instance::Instance;
     use crate::subscription::Subscriptions;
-    use rand::{distributions::Alphanumeric, Rng};
+
+    use test_utils::create_new_test_user;
 
     #[actix_rt::test]
     async fn query_notifications_does_not_return_notifications_with_unsupported_uuid() {
@@ -651,50 +652,5 @@ mod tests {
                 assert_eq!(notifications.len(), 1);
             }
         }
-    }
-
-    async fn create_new_test_user<'a, E>(executor: E) -> Result<i32, sqlx::Error>
-    where
-        E: Executor<'a>,
-    {
-        let mut transaction = executor.begin().await?;
-
-        sqlx::query!(
-            r#"
-                INSERT INTO uuid (trashed, discriminator) VALUES (0, "user")
-            "#
-        )
-        .execute(&mut transaction)
-        .await?;
-
-        let new_user_id = sqlx::query!("SELECT LAST_INSERT_ID() as id FROM uuid")
-            .fetch_one(&mut transaction)
-            .await?
-            .id as i32;
-
-        sqlx::query!(
-            r#"
-                INSERT INTO user (id, username, email, password, token)
-                VALUES (?, ?, ?, "", ?)
-            "#,
-            new_user_id,
-            random_string(10),
-            random_string(10),
-            random_string(10)
-        )
-        .execute(&mut transaction)
-        .await?;
-
-        transaction.commit().await?;
-
-        Ok(new_user_id)
-    }
-
-    fn random_string(nr: usize) -> String {
-        rand::thread_rng()
-            .sample_iter(&Alphanumeric)
-            .take(nr)
-            .map(char::from)
-            .collect()
     }
 }
