+++ test-utils/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "test-utils"
+version = "0.1.0"
+authors = ["Serlo Education e.V."]
+license = "Apache-2.0"
+edition = "2018"
+
+[lib]
+name = "test_utils"
+test = false
+doctest = false
+bench = false
+doc = false
+
+[dependencies]
+rand = "0.8.0"
+sqlx = { version = "0.5.0", features = ["chrono", "offline", "json", "mysql", "runtime-actix-native-tls"] }
