+++ test-utils/src/lib.rs
@@ -1,0 +1,47 @@
+use rand::{distributions::Alphanumeric, Rng};
+use sqlx;
+
+pub async fn create_new_test_user<'a, E>(executor: E) -> Result<i32, sqlx::Error>
+where
+    E: sqlx::Acquire<'a, Database = sqlx::MySql>,
+{
+    let mut transaction = executor.begin().await?;
+
+    sqlx::query!(
+        r#"
+                INSERT INTO uuid (trashed, discriminator) VALUES (0, "user")
+            "#
+    )
+    .execute(&mut transaction)
+    .await?;
+
+    let new_user_id = sqlx::query!("SELECT LAST_INSERT_ID() as id FROM uuid")
+        .fetch_one(&mut transaction)
+        .await?
+        .id as i32;
+
+    sqlx::query!(
+        r#"
+                INSERT INTO user (id, username, email, password, token)
+                VALUES (?, ?, ?, "", ?)
+            "#,
+        new_user_id,
+        random_string(10),
+        random_string(10),
+        random_string(10)
+    )
+    .execute(&mut transaction)
+    .await?;
+
+    transaction.commit().await?;
+
+    Ok(new_user_id)
+}
+
+fn random_string(nr: usize) -> String {
+    rand::thread_rng()
+        .sample_iter(&Alphanumeric)
+        .take(nr)
+        .map(char::from)
+        .collect()
+}
