+++ tests/user.rs
@@ -4,9 +4,8 @@ mod tests {
     use serde_json::json;
     use std::str::from_utf8;
 
-    use rand::{distributions::Alphanumeric, Rng};
-    use serlo_org_database_layer::database::Executor;
     use serlo_org_database_layer::{configure_app, create_database_pool};
+    use test_utils::create_new_test_user;
 
     #[actix_rt::test]
     async fn user_activity_by_type() {
@@ -96,51 +95,5 @@ mod tests {
                 "reason": "user with id 5 does not exist"
             }
         );
-    }
-
-    // TODO: Move this utility function into an extra crate
-    async fn create_new_test_user<'a, E>(executor: E) -> Result<i32, sqlx::Error>
-    where
-        E: Executor<'a>,
-    {
-        let mut transaction = executor.begin().await?;
-
-        sqlx::query!(
-            r#"
-                INSERT INTO uuid (trashed, discriminator) VALUES (0, "user")
-            "#
-        )
-        .execute(&mut transaction)
-        .await?;
-
-        let new_user_id = sqlx::query!("SELECT LAST_INSERT_ID() as id FROM uuid")
-            .fetch_one(&mut transaction)
-            .await?
-            .id as i32;
-
-        sqlx::query!(
-            r#"
-                INSERT INTO user (id, username, email, password, token)
-                VALUES (?, ?, ?, "", ?)
-            "#,
-            new_user_id,
-            random_string(10),
-            random_string(10),
-            random_string(10)
-        )
-        .execute(&mut transaction)
-        .await?;
-
-        transaction.commit().await?;
-
-        Ok(new_user_id)
-    }
-
-    fn random_string(nr: usize) -> String {
-        rand::thread_rng()
-            .sample_iter(&Alphanumeric)
-            .take(nr)
-            .map(char::from)
-            .collect()
     }
 }
